py/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[license]
text = "MIT"

[project]
name = "booster"
version = "0.1.0"
description = "Booster Utilities"
authors = [
    { name = "Ryan Summers", email = "ryan.summers@vertigo-designs.com" },
    { name = "Robert Jördens", email = "rj@quartiq.de" },
]
dependencies = [
    # Note: keep this in sync with Cargo.toml
    "miniconf-mqtt@git+https://github.com/quartiq/miniconf@v0.13.1#subdirectory=py/miniconf-mqtt",
]
